+++ secretsmanager/data_source_address.go
@@ -117,6 +117,7 @@ func dataSourceAddress() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -155,6 +156,11 @@ func dataSourceAddressRead(ctx context.Context, d *schema.ResourceData, m interf
 
 	fileItems := getFileItemsData(secret.Files)
 	if err = d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}