+++ secretsmanager/data_source_bank_card.go
@@ -159,6 +159,7 @@ func dataSourceBankCard() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -217,6 +218,11 @@ func dataSourceBankCardRead(ctx context.Context, d *schema.ResourceData, m inter
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}