+++ secretsmanager/data_source_contact.go
@@ -185,6 +185,7 @@ func dataSourceContact() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -247,6 +248,11 @@ func dataSourceContactRead(ctx context.Context, d *schema.ResourceData, m interf
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}