+++ secretsmanager/data_source_database_credentials.go
@@ -113,6 +113,7 @@ func dataSourceDatabaseCredentials() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -161,6 +162,11 @@ func dataSourceDatabaseCredentialsRead(ctx context.Context, d *schema.ResourceDa
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}