+++ secretsmanager/data_source_driver_license.go
@@ -163,6 +163,7 @@ func dataSourceDriverLicense() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -235,6 +236,11 @@ func dataSourceDriverLicenseRead(ctx context.Context, d *schema.ResourceData, m 
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}