+++ secretsmanager/data_source_health_insurance.go
@@ -123,6 +123,7 @@ func dataSourceHealthInsurance() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -175,6 +176,11 @@ func dataSourceHealthInsuranceRead(ctx context.Context, d *schema.ResourceData, 
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}