+++ secretsmanager/data_source_membership.go
@@ -113,6 +113,7 @@ func dataSourceMembership() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -159,6 +160,11 @@ func dataSourceMembershipRead(ctx context.Context, d *schema.ResourceData, m int
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}