+++ secretsmanager/data_source_passport.go
@@ -174,6 +174,7 @@ func dataSourcePassport() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -258,6 +259,11 @@ func dataSourcePassportRead(ctx context.Context, d *schema.ResourceData, m inter
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}