+++ secretsmanager/data_source_server_credentials.go
@@ -108,6 +108,7 @@ func dataSourceServerCredentials() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -153,6 +154,11 @@ func dataSourceServerCredentialsRead(ctx context.Context, d *schema.ResourceData
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}