+++ secretsmanager/data_source_ssh_keys.go
@@ -128,6 +128,7 @@ func dataSourceSshKeys() *schema.Resource {
 					},
 				},
 			},
+			"custom": schemaCustomFieldData(),
 		},
 	}
 }
@@ -179,6 +180,11 @@ func dataSourceSshKeysRead(ctx context.Context, d *schema.ResourceData, m interf
 
 	fileItems := getFileItemsData(secret.Files)
 	if err := d.Set("file_ref", fileItems); err != nil {
+		return diag.FromErr(err)
+	}
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
 		return diag.FromErr(err)
 	}