+++ secretsmanager/ephemeral_software_license.go
@@ -27,6 +27,7 @@ type ephemeralSoftwareLicenseModel struct {
 	ActivationDate types.String `tfsdk:"activation_date"`
 	ExpirationDate types.String `tfsdk:"expiration_date"`
 	FileRef        types.List   `tfsdk:"file_ref"`
+	Custom  types.List   `tfsdk:"custom"`
 }
 
 func NewEphemeralSoftwareLicense() ephemeral.EphemeralResource {
@@ -71,6 +72,7 @@ func (e *ephemeralSoftwareLicense) Schema(_ context.Context, _ ephemeral.SchemaR
 				Description: "Date of expiration.",
 			},
 			"file_ref": fileRefEphemeralAttribute(),
+			"custom": genericFieldEphemeralAttribute("Custom fields of the record."),
 		},
 	}
 }
@@ -130,6 +132,12 @@ func (e *ephemeralSoftwareLicense) Open(ctx context.Context, req ephemeral.OpenR
 	fileRefList, diags := fileItemsToListValue(ctx, secret.Files)
 	resp.Diagnostics.Append(diags...)
 	data.FileRef = fileRefList
+
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	customList, diags := genericFieldItemsToListValue(ctx, customItems)
+	resp.Diagnostics.Append(diags...)
+	data.Custom = customList
+
 
 	if resp.Diagnostics.HasError() {
 		return